Add playback scene
Modifies major scene 0xE's result screen and changes it to a debug menu
(minimal file load times). Upon loading a debug menu, the major scene is
checked. If it is scene 0xE, the custom text is displayed and it awaits
an EXI response. When the response is received, it advances to the match
scene.