add assemble-time optional debugging code
adds a variable named "debugFlag" which controls whether or not additional debugging code is assembled.

the additional code includes OSReport calls to notify when a desync was detected and output the original and backed up data. each frame spent waiting for the replay data is logged as well.

also migrated GetFrameIndex and GetIsFollower to be common functions.